DUOS-924 Remove VoteAPIs
