+++ src/main/java/org/broadinstitute/consent/http/resources/ConsentVoteResource.java
@@ -3,10 +3,9 @@ package org.broadinstitute.consent.http.resources;
 import com.google.inject.Inject;
 import freemarker.template.TemplateException;
 import org.broadinstitute.consent.http.models.Vote;
-import org.broadinstitute.consent.http.service.AbstractVoteAPI;
 import org.broadinstitute.consent.http.service.ElectionService;
 import org.broadinstitute.consent.http.service.EmailNotifierService;
-import org.broadinstitute.consent.http.service.VoteAPI;
+import org.broadinstitute.consent.http.service.VoteService;
 
 import javax.annotation.security.PermitAll;
 import javax.annotation.security.RolesAllowed;
@@ -30,14 +29,14 @@ import java.util.logging.Logger;
 public class ConsentVoteResource extends Resource {
 
     private final EmailNotifierService emailNotifierService;
-    private final VoteAPI api;
+    private final VoteService voteService;
     private final ElectionService electionService;
     private static final Logger logger = Logger.getLogger(ConsentVoteResource.class.getName());
 
     @Inject
-    public ConsentVoteResource(EmailNotifierService emailNotifierService, ElectionService electionService) {
+    public ConsentVoteResource(EmailNotifierService emailNotifierService, ElectionService electionService, VoteService voteService) {
         this.emailNotifierService = emailNotifierService;
-        this.api = AbstractVoteAPI.getInstance();
+        this.voteService = voteService;
         this.electionService = electionService;
     }
 
@@ -48,7 +47,7 @@ public class ConsentVoteResource extends Resource {
     public Response firstVoteUpdate(Vote rec,
                                     @PathParam("consentId") String consentId, @PathParam("id") Integer voteId){
         try {
-            Vote vote = api.updateVoteById(rec, voteId);
+            Vote vote = voteService.updateVoteById(rec, voteId);
             if(electionService.validateCollectEmailCondition(vote)){
                 try {
                     emailNotifierService.sendCollectMessage(vote.getElectionId());
@@ -70,7 +69,7 @@ public class ConsentVoteResource extends Resource {
     public Response updateConsentVote(Vote rec,
                                       @PathParam("consentId") String consentId, @PathParam("id") Integer id) {
         try {
-            Vote vote = api.updateVote(rec, id, consentId);
+            Vote vote = voteService.updateVote(rec, id, consentId);
             return Response.ok(vote).build();
         } catch (Exception e) {
             return createExceptionResponse(e);
@@ -83,7 +82,7 @@ public class ConsentVoteResource extends Resource {
     @PermitAll
     public Vote describe(@PathParam("consentId") String consentId,
                          @PathParam("id") Integer id) {
-        return api.describeVoteById(id, consentId);
+        return voteService.describeVoteById(id, consentId);
     }
 
     @DELETE
@@ -92,7 +91,7 @@ public class ConsentVoteResource extends Resource {
     @RolesAllowed(ADMIN)
     public Response deleteVote(@PathParam("consentId") String consentId, @PathParam("id") Integer id) {
         try {
-            api.deleteVote(id, consentId);
+            voteService.deleteVote(id, consentId);
             return Response.status(Response.Status.OK).entity("Vote was deleted").build();
         }  catch (Exception e) {
             return createExceptionResponse(e);
@@ -107,7 +106,7 @@ public class ConsentVoteResource extends Resource {
             if (consentId == null) {
                 return Response.status(Response.Status.BAD_REQUEST).build();
             }
-            api.deleteVotes(consentId);
+            voteService.deleteVotes(consentId);
             return Response.ok().entity("Votes for specified consent have been deleted").build();
         } catch (Exception e) {
             return createExceptionResponse(e);
