--- src/main/java/org/broadinstitute/consent/http/service/VoteAPI.java
@@ -1,27 +1,0 @@
-package org.broadinstitute.consent.http.service;
-
-import org.broadinstitute.consent.http.models.Vote;
-
-import javax.ws.rs.NotFoundException;
-import java.util.List;
-
-@Deprecated // Use VoteService
-public interface VoteAPI {
-
-    Vote updateVoteById(Vote rec,  Integer voteId) throws IllegalArgumentException;
-
-    Vote updateVote(Vote rec, Integer voteId, String referenceId) throws IllegalArgumentException;
-
-    List<Vote> describeVotes(String referenceId);
-
-    Vote describeVoteById(Integer voteId, String referenceId) throws NotFoundException;
-
-    void deleteVote(Integer voteId, String referenceId) throws IllegalArgumentException, UnknownIdentifierException;
-
-    void deleteVotes(String referenceId) throws IllegalArgumentException, UnknownIdentifierException;
-
-    List<Vote> describeVoteByTypeAndElectionId(String type, Integer electionId);
-
-    Vote describeDataOwnerVote(String requestId, Integer dataOwnerId) throws NotFoundException;
-
-}
