+++ src/main/java/org/broadinstitute/consent/http/service/VoteService.java
@@ -11,6 +11,8 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import javax.ws.rs.NotFoundException;
+
+import org.apache.commons.lang3.StringUtils;
 import org.broadinstitute.consent.http.db.DatasetAssociationDAO;
 import org.broadinstitute.consent.http.db.ElectionDAO;
 import org.broadinstitute.consent.http.db.UserDAO;
@@ -96,6 +98,26 @@ public class VoteService {
                 vote.getHasConcerns()
         );
         return voteDAO.findVoteById(vote.getVoteId());
+    }
+
+    public Vote updateVoteById(Vote rec,  Integer voteId) throws IllegalArgumentException {
+        Vote vote = voteDAO.findVoteById(voteId);
+        if (vote == null) notFoundException(voteId);
+        Integer electionId = setGeneralFields(rec, vote.getElectionId());
+        String rationale = StringUtils.isEmpty(rec.getRationale()) ? null : rec.getRationale();
+        boolean reminder = Objects.nonNull(rec.getIsReminderSent()) ? rec.getIsReminderSent() : false;
+        Date createDate = Objects.nonNull(vote.getCreateDate()) ? vote.getCreateDate() : new Date();
+        voteDAO.updateVote(rec.getVote(), rationale, new Date(), voteId, reminder, electionId, createDate, rec.getHasConcerns());
+        return voteDAO.findVoteById(voteId);
+    }
+
+    public Vote updateVote(Vote rec, Integer voteId, String referenceId) throws IllegalArgumentException {
+        if (voteDAO.checkVoteById(referenceId, voteId) == null) notFoundException(voteId);
+        Vote vote = voteDAO.findVoteById(voteId);
+        Date updateDate = rec.getVote() == null ? null : new Date();
+        String rationale = StringUtils.isNotEmpty(rec.getRationale()) ? rec.getRationale() : null;
+        voteDAO.updateVote(rec.getVote(), rationale, updateDate, voteId, false,  getElectionId(referenceId), vote.getCreateDate(), rec.getHasConcerns());
+        return voteDAO.findVoteById(voteId);
     }
 
     /**
@@ -192,6 +214,23 @@ public class VoteService {
         return mostRecentVote.orElse(votes.get(0));
     }
 
+    public List<Vote> describeVotes(String referenceId) {
+        List<Vote> resultVotes = voteDAO.findVotesByReferenceId(referenceId);
+        if (resultVotes == null || resultVotes.isEmpty()) {
+            throw new NotFoundException("Could not find vote for specified reference id. Reference id: " + referenceId);
+        }
+        return resultVotes;
+    }
+
+    public Vote describeVoteById(Integer voteId, String referenceId)
+            throws IllegalArgumentException {
+        Vote vote = voteDAO.findVoteById(voteId);
+        if (vote == null) {
+            throw new NotFoundException("Could not find vote for specified id. Vote id: " + voteId);
+        }
+        return vote;
+    }
+
     /**
      * Delete any votes in Open elections for the specified user in the specified Dac.
      *
@@ -211,6 +250,35 @@ public class VoteService {
                 voteDAO.removeVotesByIds(openUserVoteIds);
             }
         }
+    }
+
+    public void deleteVote(Integer voteId, String referenceId) {
+        if (voteDAO.checkVoteById(referenceId, voteId) == null) {
+            throw new NotFoundException("Does not exist vote for the specified id. Id: " + voteId);
+        }
+        voteDAO.deleteVoteById(voteId);
+
+    }
+
+    public void deleteVotes(String referenceId)
+            throws IllegalArgumentException, UnknownIdentifierException {
+        if (electionDAO.findElectionsWithFinalVoteByReferenceId(referenceId) == null) {
+            throw new IllegalArgumentException();
+        }
+        voteDAO.deleteVotes(referenceId);
+
+    }
+
+    public List<Vote> describeVoteByTypeAndElectionId(String type, Integer electionId) {
+        return voteDAO.findVoteByTypeAndElectionId(electionId, type);
+    }
+
+    public Vote describeDataOwnerVote(String requestId, Integer dataOwnerId) throws NotFoundException {
+        Vote vote = voteDAO.findVotesByReferenceIdTypeAndUser(requestId, dataOwnerId, VoteType.DATA_OWNER.getValue());
+        if(vote == null) {
+            throw new NotFoundException("Vote doesn't exist for the specified dataOwnerId");
+        }
+        return vote;
     }
 
     private boolean isDacChairPerson(Dac dac, User user) {
@@ -243,6 +311,25 @@ public class VoteService {
         if (Objects.isNull(voteDAO.findVoteById(vote.getVoteId()))) {
             throw new IllegalArgumentException("No vote exists with the id of " + vote.getVoteId());
         }
+    }
+
+    private Integer getElectionId(String referenceId) {
+        Integer electionId = electionDAO.getOpenElectionIdByReferenceId(referenceId);
+        if (electionId == null) {
+            throw new IllegalArgumentException("The specified object does not have an election");
+        }
+        return electionId;
+    }
+
+    private Integer setGeneralFields(Vote rec, Integer electionId) {
+        rec.setCreateDate(new Date());
+        rec.setElectionId(electionId);
+        rec.setType(rec.getType());
+        return electionId;
+    }
+
+    private void notFoundException(Integer voteId){
+        throw new NotFoundException("Could not find vote for specified id. Vote id: " + voteId);
     }
 
 }
