+++ src/test/java/org/broadinstitute/consent/http/resources/ConsentVoteResourceTest.java
@@ -4,19 +4,13 @@ package org.broadinstitute.consent.http.resources;
 import org.apache.commons.lang3.RandomUtils;
 import org.broadinstitute.consent.http.models.Consent;
 import org.broadinstitute.consent.http.models.Vote;
-import org.broadinstitute.consent.http.service.AbstractVoteAPI;
 import org.broadinstitute.consent.http.service.ElectionService;
 import org.broadinstitute.consent.http.service.EmailNotifierService;
-import org.broadinstitute.consent.http.service.VoteAPI;
+import org.broadinstitute.consent.http.service.VoteService;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.powermock.api.mockito.PowerMockito;
-import org.powermock.core.classloader.annotations.PowerMockIgnore;
-import org.powermock.core.classloader.annotations.PrepareForTest;
-import org.powermock.modules.junit4.PowerMockRunner;
 
 import javax.ws.rs.core.Response;
 import java.util.UUID;
@@ -26,16 +20,10 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 
-@SuppressWarnings("deprecation")
-@RunWith(PowerMockRunner.class)
-@PowerMockIgnore("jdk.internal.reflect.*")
-@PrepareForTest({
-        AbstractVoteAPI.class
-})
 public class ConsentVoteResourceTest {
 
     @Mock
-    private VoteAPI voteAPI;
+    private VoteService voteService;
     @Mock
     private ElectionService electionService;
     @Mock
@@ -46,12 +34,10 @@ public class ConsentVoteResourceTest {
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
-        PowerMockito.mockStatic(AbstractVoteAPI.class);
     }
 
     private void initResource() {
-        when(AbstractVoteAPI.getInstance()).thenReturn(voteAPI);
-        resource = new ConsentVoteResource(emailNotifierService, electionService);
+        resource = new ConsentVoteResource(emailNotifierService, electionService, voteService);
     }
 
     @Test
@@ -62,7 +48,7 @@ public class ConsentVoteResourceTest {
         vote.setVoteId(RandomUtils.nextInt(100, 1000));
         vote.setVote(false);
         vote.setRationale("Test");
-        when(voteAPI.updateVoteById(any(), any())).thenReturn(vote);
+        when(voteService.updateVoteById(any(), any())).thenReturn(vote);
         when(electionService.validateCollectEmailCondition(any())).thenReturn(true);
         doNothing().when(emailNotifierService).sendCollectMessage(any());
         initResource();
